Addressed the following concerns:
The rules are that the top line is always the shortest, with each following lines a little longer, with the bottom line being the longest. The text on the bottom line should always finish at least 2 centimetres away from the existing NATO or STO logo and should sit just above the rule.

The font sizes are as follows:

·         1 to 2 lines of heading text = Arial Bold 11 pt

·         3 lines of heading text = Arial Bold 10 pt

·         4 lines of text = Arial Bold 8 or 9 pt (depending on the length of the chapter title)

In terms of heading numbering, I noticed in the sample PDF provided that the NOMENCLATURE heading was not numbered (and it should be) – if this was Chapter 8, then it should be numbered “8.0”. In addition, the ACKNOWLEDGMENTS heading should also be numbered.

With regards to the page number in the footer, the dash is bigger than the one used in the STO MS Word Template (see attached).

I also noticed that there was no paragraph spacing on pages 8-3 and 8-4 – there should be 12 pts of space between each paragraph.

Also there should be NO indentation of 1st line of text in a paragraph. Examples of paragraphs, spacing and a caption can be seen in the PDF file attached.

With regards to URLs cited in text, these are always coloured NATO blue (Pantone # 128).

Finally, as I mentioned before, figure and table captions should be ‘Title’ Case and not ‘Sentence’ Case.